backend/requirements.txt
# Core
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
requests
httpx
loguru

# Data Ingestion
beautifulsoup4
lxml
pandas
numpy

# AI & Agents
langchain
langgraph
langchain-openai
openai
mcp[cli]

# Concurrency & Cache
redis
celery

# Testing
pytest
pytest-asyncio
mock
